LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.draw.space.basic-expected.txt

2d.text.draw.space.basic

U+0020 is rendered the correct size (1em wide)


PASS U+0020 is rendered the correct size (1em wide)
